Warning cleanup

+++ src/game/game.c
@@ -358,7 +358,7 @@ const struct cmdline_options_s main_cmdline_options[] = {
 
 static bool game_cfg_check_new_game_opts(void *val)
 {
-    int v2, v = (int)val;
+    int v2, v = (int)(intptr_t)val;
     v2 = v % 10;
     v = v / 10;
     if (v2 > DIFFICULTY_NUM) {
